Refactor getMeshes to aframe-utils
